+++ src/bin/daemon/LCDPlugins/LCDPlugin.hpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2021  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2022  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -139,6 +139,12 @@ class LCDPlugin
 			const bool lockedPlugin,
 			const uint16_t PBMXPos = 1,
 			const uint16_t PBMYPos = 1
+		);
+
+		void drawVerticalLineOnPBMFrame(
+			const uint16_t PBMXPos,
+			const uint16_t PBMYPos,
+			const uint16_t size
 		);
 
 	private: